更新代码

+++ discuz/cleancache.php
@@ -9,30 +9,25 @@ function cleanCacheFile($dir, $type = 1)
 {
     $handle = opendir($dir);
     while ($file = readdir($handle)) {
-        if ($file !== ".." && $file !== ".") {
+        if ($file !== ".." && $file !== "." && $file !== "attachment") {
             $f = $dir . DIRECTORY_SEPARATOR . $file;
             if (is_file($f)) {
-//                echo "|--" . $file . "<br>";
                 if ($type === 1) {
                     unlink($f);
                 }
             } else {
-                if ($type === 2) {
-                    echo "目录名：" . $file . "<br>";
-                    $myfile = $f . DIRECTORY_SEPARATOR . "index.htm";
-                    $fileRes = fopen($myfile, "w+");
+                if ($type === 1) {
+                    cleanCacheFile($f);
                 }
-//                echo "--" . $file . "<br>";
-                if ($file !== "attachment") {
-                    if ($type === 1) {
-                        cleanCacheFile($f);
-                    }
-                    if ($type === 2) {
-                        cleanCacheFile($f, 2);
-                    }
+                if ($type === 2) {
+                    cleanCacheFile($f, 2);
                 }
             }
         }
+    }
+    if ($type === 2 && is_dir($dir)) {
+        $myfile = $dir . DIRECTORY_SEPARATOR . "index.htm";
+        fopen($myfile, "w+");
     }
 }
 
@@ -42,6 +37,6 @@ echo "清理完成data目录缓存。<br>";
 //echo "清理完成uc_server/data目录缓存。<br>";
 
 
-// 往data目录，每个子目录创建一个index.htm空文件
+// 往data目录，每个子目录创建一个index.htm空文件，保留目录结构
 cleanCacheFile($path, 2);
 echo "完成在每个目录下创建空文件index.htm。";
